test: refactor to error_test.go and remove commented lines
