+++ internal/write/service_test.go
@@ -337,13 +337,10 @@ func TestMaxRetryTime(t *testing.T) {
 	b := NewBatch("2\n", exp)
 	// First batch will  be checked against maxRetryTime and it will expire. New batch will fail and it will added to retry queue
 	err = srv.HandleWrite(ctx, b)
-	//fmt.Printf("DEBUG err %v\n", err)
-	//fmt.Printf("DEBUG err %v\n", err)
 	require.NotNil(t, err)
 	// 1st Batch expires and writing 2nd trows error
 	assert.Equal(t, "write failed (attempts 1): Unexpected status code 429", err.Error())
 	assert.Equal(t, 1, srv.retryQueue.list.Len())
-	// fmt.Printf("DEBUG Header len: %d\n", len(err.(*http.Error).Header))
 
 	//wait until remaining accumulated retryDelay has passed, because there hasn't been a successful write yet
 	<-time.After(time.Until(srv.lastWriteAttempt.Add(time.Millisecond * time.Duration(srv.retryDelay))))
